fix crash when too many categories disappears from db
The problem here is that we only store a category number, so what can
happen is that repo disappears and with it potentially all it's
categories. We then want to restore to a category index that is too
large for the whole list. Instead reset back to Whats New.

Note: we'll still potentially restore the wrong category when some of
them disappear, but at least we'll not go into a crash loop.